--- native/src/main/java/io/ballerina/stdlib/ftp/util/Util.java
@@ -1,44 +1,0 @@
-/*
- * Copyright (c) 2024 WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
- *
- * WSO2 Inc. licenses this file to you under the Apache License,
- * Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package io.ballerina.stdlib.ftp.util;
-
-import io.ballerina.runtime.api.creators.ErrorCreator;
-import io.ballerina.runtime.api.values.BObject;
-
-import java.util.concurrent.CompletableFuture;
-
-import static io.ballerina.stdlib.ftp.util.FtpConstants.ENTITY_BYTE_STREAM;
-
-/**
- * Util methods for FTP operations.
- */
-public class Util {
-    public static Object getResult(CompletableFuture<Object> balFuture) {
-        try {
-            return balFuture.get();
-        } catch (Throwable throwable) {
-            throw ErrorCreator.createError(throwable);
-        }
-    }
-
-    public static void handleStreamEnd(BObject entity, BufferHolder bufferHolder) {
-        entity.addNativeData(ENTITY_BYTE_STREAM, null);
-        bufferHolder.setTerminal(true);
-    }
-}
